+++ server/routes/types.route.ts
@@ -31,12 +31,24 @@ router.post('/', async (req, res) => {
 
 // Modifica un tipo por su nombre
 router.put('/:name', async (req, res) => {
+    const { name } = req.params;
+    const typeWithName = await TypesModel.find({ name }).lean().exec();
+    if (typeWithName.length === 0) {
+        res.status(404).json({ message: `No records with ${name} name` });
+        return;
+    }
     await TypesModel.updateOne({ name: req.params.name }, { $set: { name: req.body.name } });
     res.status(202).json({ message: 'Successfully modified' });
 });
 
 // Elimina un tipo por su nombre
 router.delete('/:name', async (req, res) => {
+    const { name } = req.params;
+    const typeWithName = await TypesModel.find({ name }).lean().exec();
+    if (typeWithName.length === 0) {
+        res.status(404).json({ message: `No records with ${name} name` });
+        return;
+    }
     await TypesModel.deleteOne({ name: req.params.name });
     res.status(202).json({ message: 'Successfully deleted' });
 });
